Revert "Merge branch 'SkyblockOverhaul:Diana-V2' into Diana-V2"
This reverts commit b845d3d4c9f3a6d0b34138bbabcb100ea9550d3d, reversing
changes made to 67fd94c1c2a253d5674296dd79032dc1a1e24ba5.
