+++ gradle.properties
@@ -3,10 +3,10 @@ systemProp.javax.xml.transform.TransformerFactory=com.sun.org.apache.xalan.inter
 systemProp.javax.xml.parsers.SAXParserFactory=com.sun.org.apache.xerces.internal.jaxp.SAXParserFactoryImpl
 
 # Gradle configuration
-org.gradle.jvmargs=-Xms1m -Xmx4g -XX:+UnlockExperimentalVMOptions -XX:SoftMaxHeapSize=2g -XX:+UseStringDeduplication -XX:+UseZGC -XX:+IgnoreUnrecognizedVMOptions -XX:+UseCompactObjectHeaders
+org.gradle.jvmargs=-Xmx8192m -Xms8192m -XX:MaxMetaspaceSize=1024m -XX:+HeapDumpOnOutOfMemoryError -Dfile.encoding=UTF-8
+org.gradle.parallel.threads=12
 org.gradle.parallel=true
-
-kotlin.daemon.jvmargs=-Xmx2g
+org.gradle.daemon=false
 
 # loom
 loom.ignoreDependencyLoomVersionValidation=true
@@ -21,15 +21,13 @@ mod.version=0.1.5-beta
 mod.group=net.sbo.mod
 
 # Dependencies
-uc.version=444
-elementa.version=714
+uc.version=430
+elementa.version=710
 rconfig.version.1.21.5=3.5.13
 rconfig.version.1.21.7=3.7.2
-rconfig.version.1.21.10=3.9.1
 hmapi.version=1.0.1+1.21.2
 modmenu.version.1.21.5=14.0.0-rc.2
 modmenu.version.1.21.7=15.0.0-beta.3
-modmenu.version.1.21.10=16.0.0-rc.1
 vexel.version=100
 devauth.version=1.2.1
 autoservice.version=1.1.0
