+++ root.gradle.kts
@@ -3,9 +3,7 @@ plugins {
 }
 
 preprocess {
-    "1.21.10-fabric"(1_21_10, "yarn") {
-        "1.21.7-fabric"(1_21_08, "yarn") {
-            "1.21.5-fabric"(1_21_05, "yarn")
-        }
+    "1.21.7-fabric"(1_21_08, "yarn") {
+        "1.21.5-fabric"(1_21_05, "yarn")
     }
 }
