+++ src/main/java/net/sbo/mod/mixin/accessor/BeaconBlockEntityRendererInvoker.java
@@ -1,0 +1,15 @@
+package net.sbo.mod.mixin.accessor;
+
+import net.minecraft.client.render.VertexConsumerProvider;
+import net.minecraft.client.render.block.entity.BeaconBlockEntityRenderer;
+import net.minecraft.client.util.math.MatrixStack;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Invoker;
+
+@Mixin(BeaconBlockEntityRenderer.class)
+public interface BeaconBlockEntityRendererInvoker {
+    @Invoker("renderBeam")
+    static void renderBeam(MatrixStack matrices, VertexConsumerProvider vertexConsumers, float tickDelta, float scale, long worldTime, int yOffset, int maxY, int color) {
+        throw new UnsupportedOperationException();
+    }
+}
