+++ src/main/kotlin/net/sbo/mod/utils/SoundHandler.kt
@@ -38,9 +38,7 @@ object SoundHandler {
 
     private fun currentResourcePackFormat(): Int {
         return try {
-            //#if MC >= 1.21.9
-            //$$ (SharedConstants.getGameVersion().packVersion(ResourceType.CLIENT_RESOURCES) as Int)
-            //#elseif MC >= 1.21.7
+            //#if MC >= 1.21.7
             //$$ SharedConstants.getGameVersion().packVersion(ResourceType.CLIENT_RESOURCES)
             //#else
             SharedConstants.getGameVersion().getResourceVersion(ResourceType.CLIENT_RESOURCES)
