+++ src/main/kotlin/net/sbo/mod/utils/render/RenderUtils3D.kt
@@ -6,16 +6,20 @@ import net.minecraft.client.render.*
 import net.minecraft.util.math.RotationAxis
 import net.minecraft.util.math.Vec3d
 import net.sbo.mod.SBOKotlin.mc
+import net.sbo.mod.utils.waypoint.WaypointManager
+import net.sbo.mod.mixin.accessor.BeaconBlockEntityRendererInvoker
 import net.sbo.mod.settings.categories.Customization
 import net.sbo.mod.utils.math.SboVec
 import java.awt.Color
 import kotlin.math.max
-import net.minecraft.client.util.math.MatrixStack
-import net.minecraft.client.render.VertexConsumerProvider
-import net.minecraft.util.math.MathHelper
-import net.minecraft.util.math.ColorHelper
 
-object RenderUtils3D {
+object WaypointRenderer : WorldRenderEvents.AfterTranslucent {
+    override fun afterTranslucent(context: WorldRenderContext) {
+        WaypointManager.renderAllWaypoints(context)
+    }
+}
+
+object RenderUtil {
     fun renderWaypoint(
         context: WorldRenderContext,
         text: String,
@@ -54,8 +58,8 @@ object RenderUtils3D {
             renderBeaconBeam(
                 context,
                 pos,
-                colorComponents,
-                true
+                1,
+                colorComponents
             )
 
         }
@@ -187,6 +191,47 @@ object RenderUtils3D {
     }
 
     /**
+     * Renders a beacon beam at the given location.
+     * @param context The world render context.
+     * @param pos The position in the world where the beacon beam should be rendered.
+     * @param colorComponents The RGB color components as a FloatArray (0.0 to
+     */
+    fun renderBeaconBeam( // todo: add through walls option
+        context: WorldRenderContext,
+        pos: SboVec,
+        yOffset: Int,
+        colorComponents: FloatArray,
+    ) {
+        val matrices = context.matrixStack()!!
+        val cameraPos = context.camera().pos
+        val world = context.world()
+
+        matrices.push()
+        matrices.translate(pos.x - cameraPos.x, pos.y - cameraPos.y, pos.z - cameraPos.z)
+
+        val consumers = context.consumers()!!
+        val partialTicks = context.tickCounter().getTickProgress(true)
+        val worldAge = world.time
+
+        val beamHeight = context.world().height
+        val beamColor = floatArrayOf(colorComponents[0], colorComponents[1], colorComponents[2], 1.0f)
+
+
+        BeaconBlockEntityRendererInvoker.renderBeam(
+            matrices,
+            consumers,
+            partialTicks,
+            1.0f,
+            worldAge,
+            yOffset,
+            beamHeight,
+            Color(beamColor[0], beamColor[1], beamColor[2]).rgb
+        )
+
+        matrices.pop()
+    }
+
+    /**
      * Draws a line from the player's eyes to a target point in the world.
      * @param context The world render context.
      * @param target The target position in the world.
@@ -239,165 +284,5 @@ object RenderUtils3D {
             .color(color[0], color[1], color[2], alpha)
 
         matrices.pop()
-    }
-    /**
-     * Renders a beacon beam at the given location.
-     * @param context The world render context.
-     * @param vec The position in the world where the beacon beam should be rendered.
-     * @param colorComponents The RGB color components as a FloatArray (0.0 to
-     * @param phase Whether the beam should render through walls.
-     */
-    @JvmOverloads
-    fun renderBeaconBeam(
-        ctx: WorldRenderContext,
-        vec: SboVec,
-        colorComponents: FloatArray,
-        phase: Boolean = false
-    ) {
-        val consumers = ctx.consumers()
-        val matrices = ctx.matrixStack() ?: return
-        val wolrd = mc.world ?: return
-        val partialTicks = ctx.tickCounter().getTickProgress(true)
-        val cam = ctx.camera().pos
-        val beamColor = floatArrayOf(colorComponents[0], colorComponents[1], colorComponents[2], 1.0f)
-
-        matrices.pushPop {
-            matrices.translate(vec.x - cam.x, (vec.y + 1.0) - cam.y, vec.z - cam.z)
-
-            renderBeam(
-                matrices,
-                consumers!!,
-                partialTicks,
-                wolrd.time,
-                Color(beamColor[0], beamColor[1], beamColor[2]).rgb,
-                phase
-            )
-        }
-    }
-
-    private fun renderBeam(
-        matricies: MatrixStack,
-        vertices: VertexConsumerProvider,
-        partialTicks: Float,
-        worldTime: Long,
-        color: Int,
-        phase: Boolean = false
-    ) {
-        val opaqueLayyer = if (phase) SboRenderLayers.BEACON_BEAM_OPAQUE_THROUGH_WALLS else SboRenderLayers.BEACON_BEAM_OPAQUE
-        val transluscentLayer = if (phase) SboRenderLayers.BEACON_BEAM_TRANSLUCENT_THROUGH_WALLS else SboRenderLayers.BEACON_BEAM_TRANSLUCENT
-        val heightScale = 1f
-        val height = 320
-        val innerRadius = 0.2f
-        val outerRadius = 0.25f
-        val time = Math.floorMod(worldTime, 40) + partialTicks
-        val fixedTime = -time
-        val wavePhase = MathHelper.fractionalPart(fixedTime * 0.2f - MathHelper.floor(fixedTime * 0.1f).toFloat())
-        val animationStep = -1f + wavePhase
-        var renderYOffest = height.toFloat() * heightScale * (0.5f / innerRadius) + animationStep
-
-        matricies.pushPop {
-            matricies.translate(0.5, 0.0, 0.5)
-
-            matricies.pushPop {
-                matricies.multiply(RotationAxis.POSITIVE_Y.rotationDegrees(time * 2.25f - 45f))
-
-                renderBeamLayer(
-                    matricies,
-                    vertices.getBuffer(opaqueLayyer),
-                    color,
-                    0f,
-                    innerRadius,
-                    innerRadius,
-                    0f,
-                    -innerRadius,
-                    0f,
-                    0f,
-                    -innerRadius,
-                    renderYOffest,
-                    animationStep
-                )
-            }
-
-            renderYOffest = height.toFloat() * heightScale + animationStep
-
-            renderBeamLayer(
-                matricies,
-                vertices.getBuffer(transluscentLayer),
-                ColorHelper.withAlpha(32, color),
-                -outerRadius,
-                -outerRadius,
-                outerRadius,
-                -outerRadius,
-                -outerRadius,
-                outerRadius,
-                outerRadius,
-                outerRadius,
-                renderYOffest,
-                animationStep
-            )
-        }
-    }
-
-    private fun renderBeamLayer(
-        matricies: MatrixStack,
-        vertices: VertexConsumer,
-        color: Int,
-        x1: Float,
-        z1: Float,
-        x2: Float,
-        z2: Float,
-        x3: Float,
-        z3: Float,
-        x4: Float,
-        z4: Float,
-        v1: Float,
-        v2: Float
-    ) {
-        val entry = matricies.peek()
-        renderBeamFace(entry, vertices, color, x1, z1, x2, z2, v1, v2)
-        renderBeamFace(entry, vertices, color, x4, z4, x3, z3, v1, v2)
-        renderBeamFace(entry, vertices, color, x2, z2, x4, z4, v1, v2)
-        renderBeamFace(entry, vertices, color, x3, z3, x1, z1, v1, v2)
-    }
-
-    private fun renderBeamFace(
-        matrix: MatrixStack.Entry,
-        vertices: VertexConsumer,
-        color: Int,
-        x1: Float,
-        z1: Float,
-        x2: Float,
-        z2: Float,
-        v1: Float,
-        v2: Float
-    ) {
-        renderBeamVertex(matrix, vertices, color, 320, x1, z1, 1f, v1)
-        renderBeamVertex(matrix, vertices, color, 0, x1, z1, 1f, v2)
-        renderBeamVertex(matrix, vertices, color, 0, x2, z2, 0f, v2)
-        renderBeamVertex(matrix, vertices, color, 320, x2, z2, 0f, v1)
-    }
-
-    private fun renderBeamVertex(
-        matrix: MatrixStack.Entry,
-        vertices: VertexConsumer,
-        color: Int,
-        y: Int,
-        x: Float,
-        z: Float,
-        u: Float,
-        v: Float
-    ) {
-        vertices
-            .vertex(matrix, x, y.toFloat(), z)
-            .color(color)
-            .texture(u, v)
-            .overlay(OverlayTexture.DEFAULT_UV)
-            .light(15728880).normal(matrix, 0f, 1f, 0f)
-    }
-
-    private fun MatrixStack.pushPop(function: () -> Unit) {
-        this.push()
-        function()
-        this.pop()
     }
 }
