+++ src/main/kotlin/net/sbo/mod/utils/render/SboRenderLayers.kt
@@ -1,11 +1,49 @@
 package net.sbo.mod.utils.render
 
+import com.mojang.blaze3d.pipeline.BlendFunction
+import com.mojang.blaze3d.pipeline.RenderPipeline
+import com.mojang.blaze3d.platform.DepthTestFunction;
+import com.mojang.blaze3d.vertex.VertexFormat.DrawMode;
+
 import net.minecraft.client.gl.RenderPipelines
 import net.minecraft.client.render.RenderLayer
 import net.minecraft.client.render.RenderPhase
-import net.minecraft.client.render.block.entity.BeaconBlockEntityRenderer
-import net.minecraft.util.TriState
+import net.minecraft.client.render.VertexFormats;
+import net.minecraft.util.Identifier;
 import java.util.OptionalDouble
+
+object SboRenderPipelines {
+    val FILLED_BOX_THROUGH_WALLS: RenderPipeline = RenderPipelines.register(
+        RenderPipeline.builder(RenderPipelines.POSITION_COLOR_SNIPPET)
+            .withLocation(Identifier.of("sbo", "pipeline/debug_filled_box_through_walls"))
+            .withVertexFormat(VertexFormats.POSITION_COLOR, DrawMode.TRIANGLE_STRIP)
+            .withDepthTestFunction(DepthTestFunction.NO_DEPTH_TEST)
+            .build()
+    )
+
+    val LINES: RenderPipeline = RenderPipelines.register(
+        RenderPipeline.builder(*arrayOf<RenderPipeline.Snippet?>(RenderPipelines.RENDERTYPE_LINES_SNIPPET))
+            .withLocation(Identifier.of("sbo", "pipeline/line_strip"))
+            .withVertexFormat(VertexFormats.POSITION_COLOR_NORMAL, DrawMode.LINES)
+            .withCull(false)
+            .withBlend(BlendFunction.TRANSLUCENT)
+            .withDepthWrite(true)
+            .withDepthTestFunction(DepthTestFunction.LEQUAL_DEPTH_TEST)
+            .build()
+    )
+
+    val LINES_THROUGH_WALLS: RenderPipeline = RenderPipelines.register(
+        RenderPipeline.builder(*arrayOf<RenderPipeline.Snippet?>(RenderPipelines.RENDERTYPE_LINES_SNIPPET))
+            .withLocation(Identifier.of("sbo", "pipeline/line_through_walls"))
+            .withShaderDefine("shad")
+            .withVertexFormat(VertexFormats.POSITION_COLOR_NORMAL, DrawMode.LINES)
+            .withCull(false)
+            .withBlend(BlendFunction.TRANSLUCENT)
+            .withDepthWrite(false)
+            .withDepthTestFunction(DepthTestFunction.NO_DEPTH_TEST)
+            .build()
+    )
+}
 
 object SboRenderLayers {
     @JvmField
@@ -59,80 +97,4 @@ object SboRenderLayers {
             )
         }
     }
-
-    val BEACON_BEAM_OPAQUE: RenderLayer = RenderLayer.of(
-        "beacon_beam_opaque",
-        1536,
-        false,
-        true,
-        SboRenderPipelines.BEACON_BEAM_OPAQUE,
-        RenderLayer.MultiPhaseParameters.builder()
-            .texture(
-                RenderPhase.Texture(
-                    BeaconBlockEntityRenderer.BEAM_TEXTURE,
-                    //#if MC < 1.21.7
-                    TriState.FALSE,
-                    //#endif
-                    false
-                )
-            )
-            .build(false)
-    )
-
-    val BEACON_BEAM_OPAQUE_THROUGH_WALLS: RenderLayer = RenderLayer.of(
-        "beacon_beam_opaque_through_walls",
-        1536,
-        false,
-        true,
-        SboRenderPipelines.BEACON_BEAM_OPAQUE_THROUGH_WALLS,
-        RenderLayer.MultiPhaseParameters.builder()
-            .texture(
-                RenderPhase.Texture(
-                    BeaconBlockEntityRenderer.BEAM_TEXTURE,
-                    //#if MC < 1.21.7
-                    TriState.FALSE,
-                    //#endif
-                    false
-                )
-            )
-            .build(false)
-    )
-
-    val BEACON_BEAM_TRANSLUCENT: RenderLayer = RenderLayer.of(
-        "beacon_beam_translucent",
-        1536,
-        false,
-        true,
-        SboRenderPipelines.BEACON_BEAM_TRANSLUCENT,
-        RenderLayer.MultiPhaseParameters.builder()
-            .texture(
-                RenderPhase.Texture(
-                    BeaconBlockEntityRenderer.BEAM_TEXTURE,
-                    //#if MC < 1.21.7
-                    TriState.FALSE,
-                    //#endif
-                    false
-                )
-            )
-            .build(false)
-    )
-
-    val BEACON_BEAM_TRANSLUCENT_THROUGH_WALLS: RenderLayer = RenderLayer.of(
-        "devonian_beacon_beam_translucent_esp",
-        1536,
-        false,
-        true,
-        SboRenderPipelines.BEACON_BEAM_TRANSLUCENT_THROUGH_WALLS,
-        RenderLayer.MultiPhaseParameters.builder()
-            .texture(
-                RenderPhase.Texture(
-                    BeaconBlockEntityRenderer.BEAM_TEXTURE,
-                    //#if MC < 1.21.7
-                    TriState.FALSE,
-                    //#endif
-                    false
-                )
-            )
-            .build(false)
-    )
 }
