--- src/main/kotlin/net/sbo/mod/utils/render/SboRenderPipelines.kt
@@ -1,65 +1,0 @@
-package net.sbo.mod.utils.render
-
-import com.mojang.blaze3d.pipeline.BlendFunction
-import com.mojang.blaze3d.pipeline.RenderPipeline
-import com.mojang.blaze3d.platform.DepthTestFunction
-import com.mojang.blaze3d.vertex.VertexFormat.DrawMode
-import net.minecraft.client.gl.RenderPipelines
-import net.minecraft.client.render.VertexFormats
-import net.minecraft.util.Identifier
-
-object SboRenderPipelines {
-    val FILLED_BOX_THROUGH_WALLS: RenderPipeline = RenderPipelines.register(
-        RenderPipeline.builder(RenderPipelines.POSITION_COLOR_SNIPPET)
-            .withLocation(Identifier.of("sbo", "pipeline/debug_filled_box_through_walls"))
-            .withVertexFormat(VertexFormats.POSITION_COLOR, DrawMode.TRIANGLE_STRIP)
-            .withDepthTestFunction(DepthTestFunction.NO_DEPTH_TEST)
-            .build()
-    )
-
-    val LINES: RenderPipeline = RenderPipelines.register(
-        RenderPipeline.builder(*arrayOf<RenderPipeline.Snippet?>(RenderPipelines.RENDERTYPE_LINES_SNIPPET))
-            .withLocation(Identifier.of("sbo", "pipeline/line_strip"))
-            .withVertexFormat(VertexFormats.POSITION_COLOR_NORMAL, DrawMode.LINES)
-            .withCull(false)
-            .withBlend(BlendFunction.TRANSLUCENT)
-            .withDepthWrite(true)
-            .withDepthTestFunction(DepthTestFunction.LEQUAL_DEPTH_TEST)
-            .build()
-    )
-
-    val LINES_THROUGH_WALLS: RenderPipeline = RenderPipelines.register(
-        RenderPipeline.builder(*arrayOf<RenderPipeline.Snippet?>(RenderPipelines.RENDERTYPE_LINES_SNIPPET))
-            .withLocation(Identifier.of("sbo", "pipeline/line_through_walls"))
-            .withShaderDefine("shad")
-            .withVertexFormat(VertexFormats.POSITION_COLOR_NORMAL, DrawMode.LINES)
-            .withCull(false)
-            .withBlend(BlendFunction.TRANSLUCENT)
-            .withDepthWrite(false)
-            .withDepthTestFunction(DepthTestFunction.NO_DEPTH_TEST)
-            .build()
-    )
-
-    val BEACON_BEAM_OPAQUE: RenderPipeline = RenderPipeline.builder(RenderPipelines.RENDERTYPE_BEACON_BEAM_SNIPPET)
-        .withLocation("sbo-kotlin/beacon_beam_opaque")
-        .build()
-
-    val BEACON_BEAM_OPAQUE_THROUGH_WALLS: RenderPipeline = RenderPipeline.builder(RenderPipelines.RENDERTYPE_BEACON_BEAM_SNIPPET)
-        .withLocation("sbo-kotlin/beacon_beam_opaque_through_walls")
-        .withDepthWrite(false)
-        .withDepthTestFunction(DepthTestFunction.NO_DEPTH_TEST)
-        .build()
-
-    val BEACON_BEAM_TRANSLUCENT: RenderPipeline = RenderPipeline.builder(RenderPipelines.RENDERTYPE_BEACON_BEAM_SNIPPET)
-        .withLocation("sbo-kotlin/beacon_beam_translucent")
-        .withDepthWrite(false)
-        .withBlend(BlendFunction.TRANSLUCENT)
-        .build()
-
-    val BEACON_BEAM_TRANSLUCENT_THROUGH_WALLS: RenderPipeline = RenderPipeline.builder(RenderPipelines.RENDERTYPE_BEACON_BEAM_SNIPPET)
-        .withLocation("sbo-kotlin/beacon_beam_translucent_through_walls")
-        .withDepthWrite(false)
-        .withBlend(BlendFunction.TRANSLUCENT)
-        .withDepthTestFunction(DepthTestFunction.NO_DEPTH_TEST)
-        .build()
-}
