+++ src/main/kotlin/net/sbo/mod/utils/waypoint/WaypointManager.kt
@@ -31,9 +31,6 @@ import kotlin.collections.iterator
 import kotlin.math.cos
 import kotlin.math.roundToInt
 import kotlin.text.get
-//#if MC >= 1.21.9
-//$$ import net.fabricmc.fabric.api.client.rendering.v1.world.WorldRenderEvents
-//#endif
 
 object WaypointManager {
     var guessWp: Waypoint? = null
@@ -192,11 +189,7 @@ object WaypointManager {
             focusedGuess = bestGuess.takeIf { maxDot >= cosThreshold }
         }
 
-        //#if MC >= 1.21.9
-        //$$ WorldRenderEvents.BEFORE_TRANSLUCENT.register(WaypointRenderer)
-        //#else
         WorldRenderEvents.AFTER_TRANSLUCENT.register(WaypointRenderer)
-        //#endif
     }
 
     @SboEvent
