+++ src/main/resources/fabric.mod.json
@@ -44,5 +44,8 @@
 		"java": ">=21",
 		"fabric-api": "*",
 		"fabric-language-kotlin": "*"
+	},
+	"suggests": {
+		"another-mod": "*"
 	}
 }
