+++ src/main/resources/mixins.sbo-kotlin.json
@@ -10,6 +10,7 @@
     "PlayerInteractMixin",
     "ScreenHandlerSlotUpdateS2CPacketMixin",
     "ScreenMixin",
+    "accessor.BeaconBlockEntityRendererInvoker",
     "EntityMixin"
   ],
   "injectors": {
