--- versions/1.21.10-fabric/.gitignore
@@ -1,7 +1,0 @@
-# Ignore everything in this directory
-*
-# Except this file
-!.gitignore
-!gradle.properties
-!src/**
-!src/
