--- versions/1.21.10-fabric/gradle.properties
@@ -1,1 +1,0 @@
-essential.loom.disableUnpick=true
