--- versions/1.21.10-fabric/src/main/resources/mixins.sbo-kotlin.json
@@ -1,18 +1,0 @@
-{
-  "required": true,
-  "package": "net.sbo.mod.mixin",
-  "compatibilityLevel": "JAVA_21",
-  "client": [
-    "GameMenuScreenMixin",
-    "GameRenderMixin",
-    "MinecraftClientMixin",
-    "PacketMixin",
-    "PlayerInteractMixin",
-    "ScreenHandlerSlotUpdateS2CPacketMixin",
-    "ScreenMixin",
-    "EntityMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
