--- versions/1.21.5-fabric/.gitignore
@@ -1,6 +1,0 @@
-# Ignore everything in this directory
-*
-# Except this file
-!.gitignore
-!src/**
-!src/
